UIDAI 2026/ALEIS/requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
PyYAML
